pyproject.toml
[project]
name = "song-shake"
version = "0.11.0"
description = "Enrich YouTube Music playlists with AI-powered genre, mood, instrument, and BPM metadata using Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
    "structlog>=24.0.0",
    "tinydb>=4.8.2",
    "typer>=0.21.1",
    "ytmusicapi>=1.11.5",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "requests>=2.0.0",
    "pyjwt>=2.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
firebase = [
    "firebase-admin>=6.0.0",
]

[project.scripts]
song-shake = "song_shake.main:app"

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pillow>=12.1.1",
]
